pyproject.toml
[project]
name = "wlddc"
version = "0.2.0"
description = "Wayland monitor control MQTT agent for Home Assistant"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiomqtt>=2.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0",
    "typer>=0.9.0",
]

[project.scripts]
wlddc = "wlddc.__main__:app"

[project.optional-dependencies]
dev = ["pytest>=7.0.0", "pytest-asyncio>=0.21.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = true
